requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
browser-use==0.11.2
google-genai==1.56.0
passporteye==2.2.1
pymupdf==1.24.9
pillow==10.4.0
pytesseract==0.3.13
